Conditional residuals (CWRES) plotted against independent variable (IDV)

Plot of conditional residuals (CWRES) plotted against independent variable (IDV).

cwres_vs_idv(xpdb, aes = NULL, group = "ID", type = "pls",
  facets = NULL, title = NULL, subtitle = NULL, caption = NULL,
  log = NULL, guides = TRUE, ...)

Arguments

xpdb
An xpose database object.

aes
The xpose aesthetics (eg. point_color).

group
Grouping variable to be used for lines.

type
String setting the type of plot to be used points 'p', line 'l' and smooth 's' or any combination of the three.

facets
Either a character string to use facet_wrap or a formula to use facet_grid.

title
Main title of the plot. If NULL automated title will be generated. Use FALSE to remove title and subtitle.

subtitle
the plot subtitle. If NULL automated subtitle will be generated. Use FALSE to remove subtitle.

caption
page caption. If NULL automated caption will be generated. Use FALSE to remove caption.

log
String assigning logarithmic scale to axes, can be either '', 'x', y' or 'xy'.

guides
Enable guides display (e.g. unity line).

...
Any additional aesthetics to be passed on xplot_scatter.

Value

An xpose_plot

Template titles

Template titles can be used to create highly informative diagnostics plots. They can be applied to any plot title, subtitle and caption. Template titles are defined via a single string containing key variables staring with a @ (e.g. @ofv) which will be replaced by their actual value when rendering the plot. For example '@run, @nobs observations in @nind subjects' would become 'run001, 1022 observations in 74 subjects'. The available key variables are listed under template_titles.

See also

xplot_scatter

Examples

not_run({
  cwres_vs_idv(xpdb_ex_pk)
})