Fetch data

Main internal function to get the data from different source and prepare it for plotting. Arguments are usually provided by `data_opt()`.

fetch_data(
  xpdb,
  .problem = NULL,
  .subprob = NULL,
  .method = NULL,
  .source = "data",
  simtab = FALSE,
  filter = NULL,
  tidy = FALSE,
  index_col = NULL,
  value_col = NULL,
  post_processing = NULL,
  quiet = FALSE
)

Arguments

xpdb

An xpose database object.

.problem

The problem to be used, by default returns the last one.

.subprob

The subproblem to be used, by default returns the last one.

.method

The estimation method to be used, by default returns the last one.

.source

Define the location of the data in the xpdb. Should be either 'data' to use the output tables or the name of an output file attached to the xpdb.

simtab

Only used when 'data' is defined as the source and `.problem` is default. Should the data be coming from an estimation or a simulation table.

filter

A function used to filter the data e.g. filter = function(x) x[x$TIME > 20, ] where x is the data.

tidy

Logical, whether the data should be transformed to tidy data.

index_col

Only used when 'tidy' is defined a TRUE and value_col is NULL. Column names to use as index when tidying the data.

value_col

Only used when 'tidy' is defined a TRUE and index_col is NULL. Column names to be stacked when tidying the data.

post_processing

A function used to modify the data after it has been tidied up e.g. post_processing = function(x) dplyr::mutate(.data = x, variable = as.factor(.$variable)) where x is the tidy data.

quiet

Should messages be displayed to the console.

Value

A tibble