Access model output table data

Access model output table data from an xpdb object.

get_data(xpdb, table = NULL, .problem = NULL, quiet)

Arguments

xpdb
An xpose_data object from which the model output file data will be extracted.

table
Name of the output table to be extracted from the xpdb e.g. 'sdtab001'. Alternative to the `.problem` argument.

.problem
Accesses all tables from the specified problem. Alternative to the `table` argument.

quiet
Logical, if FALSE messages are printed to the console.

Value

By default returns data from the last estimation problem. If only simulation problems are present then the data from last simulation will be returned instead. Object returned as tibble for single tables/problems or a named list for multiple tables/problems.

See also

list_data, xpose_data, read_nm_tables

Examples

# By table name
sdtab <- get_data(xpdb_ex_pk, 'sdtab001')
sdtab

#> # A tibble: 550 x 18
#>        ID  DOSE   AMT    SS    II  TIME   TAD   IPRED     CWRES   CPRED
#>    <fctr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>     <dbl>   <dbl>
#>  1    110   200   100     0     0   0.0   0.0 0.00000  0.000000 0.00000
#>  2    110   200     0     0     0   1.0   1.0 0.21399 -2.832600 0.58354
#>  3    110   200     0     0     0   1.5   1.5 0.29597 -1.622600 0.73677
#>  4    110   200     0     0     0   2.0   2.0 0.34878  1.098300 0.79007
#>  5    110   200     0     0     0   4.0   4.0 0.39238  0.585330 0.58843
#>  6    110   200     0     0     0   6.0   6.0 0.32855 -0.026181 0.30402
#>  7    110   200     0     0     0   8.0   8.0 0.24921  0.093976 0.12732
#>  8    112   200   100     0     0   0.0   0.0 0.00000  0.000000 0.00000
#>  9    112   200     0     0     0   0.5   0.5 0.55580  0.034146 0.62731
#> 10    112   200     0     0     0   1.0   1.0 0.71162 -0.416810 0.78763
#> # ... with 540 more rows, and 8 more variables: IWRES <dbl>, EVID <dbl>,
#> #   A1 <dbl>, A2 <dbl>, DV <dbl>, PRED <dbl>, RES <dbl>, WRES <dbl>

# By problem
tables <- get_data(xpdb_ex_pk, .problem = 1)
tables

#> # A tibble: 550 x 31
#>        ID    SEX   MED1   MED2  DOSE   AMT    SS    II  TIME   TAD   IPRED
#>    <fctr> <fctr> <fctr> <fctr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>
#>  1    110      1      0      1   200   100     0     0   0.0   0.0 0.00000
#>  2    110      1      0      1   200     0     0     0   1.0   1.0 0.21399
#>  3    110      1      0      1   200     0     0     0   1.5   1.5 0.29597
#>  4    110      1      0      1   200     0     0     0   2.0   2.0 0.34878
#>  5    110      1      0      1   200     0     0     0   4.0   4.0 0.39238
#>  6    110      1      0      1   200     0     0     0   6.0   6.0 0.32855
#>  7    110      1      0      1   200     0     0     0   8.0   8.0 0.24921
#>  8    112      1      1      1   200   100     0     0   0.0   0.0 0.00000
#>  9    112      1      1      1   200     0     0     0   0.5   0.5 0.55580
#> 10    112      1      1      1   200     0     0     0   1.0   1.0 0.71162
#> # ... with 540 more rows, and 20 more variables: CWRES <dbl>, CPRED <dbl>,
#> #   IWRES <dbl>, EVID <dbl>, A1 <dbl>, A2 <dbl>, DV <dbl>, PRED <dbl>,
#> #   RES <dbl>, WRES <dbl>, CLCR <dbl>, AGE <dbl>, WT <dbl>, KA <dbl>, CL <dbl>,
#> #   V <dbl>, ALAG1 <dbl>, ETA1 <dbl>, ETA2 <dbl>, ETA3 <dbl>

# Tip to list available tables in the xpdb
print(xpdb_ex_pk)

#> run001.lst overview: 
#>  - Software: nonmem 7.3.0 
#>  - Attached files (memory usage 1.3 Mb): 
#>    + obs tabs: $prob no.1: catab001.csv, cotab001, patab001, sdtab001 
#>    + sim tabs: $prob no.2: simtab001.zip 
#>    + output files: run001.cor, run001.cov, run001.ext, run001.grd, run001.phi, run001.shk 
#>    + special: <none> 
#>  - gg_theme: theme_readable 
#>  - xp_theme: theme_xp_default 
#>  - Options: dir = analysis/models/pk/, quiet = FALSE, manual_import = NULL