Access model summary data

Access model summary data from an xpdb object.

get_summary(xpdb, problem = NULL, subprob = NULL, only_last = FALSE)

Arguments

xpdb
An xpose_data object from which the summary data will be extracted.

problem
The problem to be used, by default returns the last one for each label.

subprob
The subproblem to be used, by default returns the last one for each label.

only_last
Logical, if TRUE only the last record for each label is returned in case of multiple problem and/or subproblem. If FALSE all values are returned.

Value

A tibble of model summary.

See also

xpose_data, template_titles

Examples

run_summary <- get_summary(xpdb_ex_pk)
run_summary

#> # A tibble: 48 x 5
#>    problem subprob               descr       label
#>      <dbl>   <dbl>               <chr>       <chr>
#>  1       0       0     Run description       descr
#>  2       0       0       Run directory         dir
#>  3       0       0          Run errors      errors
#>  4       0       0 ESAMPLE seed number esampleseed
#>  5       0       0            Run file        file
#>  6       0       0   Number of ESAMPLE    nesample
#>  7       0       0     Reference model         ref
#>  8       0       0          Run number         run
#>  9       0       0            Software    software
#> 10       0       0      Run start time   timestart
#> # ... with 38 more rows, and 1 more variables: value <chr>