List available variables

Function listing all available variables in an xpdb object.

list_vars(xpdb, problem = NULL)

Arguments

xpdb
An xpose_data object from which the model code will be extracted.

problem
The problem to be used, by lists all available problems.

See also

set_var_types

Examples

list_vars(xpdb_ex_pk)

#> 
#> List of available variables for problem no. 1 
#>  - Subject identifier (id)               : ID
#>  - Dependent variable (dv)               : DV
#>  - Independent variable (idv)            : TAD
#>  - Occasion flag (occ)                   : OCC
#>  - Dose amount (amt)                     : AMT
#>  - Event identifier (evid)               : EVID
#>  - Model typical predictions (pred)      : PRED
#>  - Model individual predictions (ipred)  : IPRED
#>  - Model parameter (param)               : KA, CL, V, ALAG1
#>  - Eta (eta)                             : ETA1, ETA2, ETA3
#>  - Residuals (res)                       : CWRES, IWRES, RES, WRES
#>  - Categorical covariates (catcov)       : SEX, MED1, MED2
#>  - Continuous covariates (contcov)       : CLCR, AGE, WT
#>  - Compartment amounts (a)               : A1, A2
#>  - Not attributed (na)                   : DOSE, SS, II, TIME, CPRED
#> 
#> List of available variables for problem no. 2 
#>  - Subject identifier (id)               : ID
#>  - Dependent variable (dv)               : DV
#>  - Independent variable (idv)            : TAD
#>  - Occasion flag (occ)                   : OCC
#>  - Dose amount (amt)                     : AMT
#>  - Event identifier (evid)               : EVID
#>  - Model individual predictions (ipred)  : IPRED
#>  - Not attributed (na)                   : DOSE, TIME, SEX, CLCR, AGE, WT