Manually define nonmem tables to be imported

Manually provide names of the table files to be imported by xpose_data.

manual_nm_import(tab_names = c("sdtab", "mutab", "patab", "catab", "cotab",
  "mytab", "extra", "xptab", "cwtab"), tab_suffix = "", sim_suffix = "sim")

Arguments

tab_names
Provide the name of the tables to import e.g. 'sdtab', 'patab', 'cotab', 'catab' for NONMEM.

tab_suffix
Default is '', but can be changed to any character string to be used as suffix in the table names.

sim_suffix
Default is 'sim', but can be changed to any character string to be used as suffix in the simulation table names e.g. sdtab001sim.

See also

xpose_data

Examples

# NOT RUN {
# Import all names specified by default as in xpose4
xpose_data(runno = '001', manual_import = manual_nm_import())

# Import a specific table name
xpose_data(runno = '001', manual_import = manual_nm_import(tab_names = 'mytab'))
# }