Parse keywords in string based on values contained in an xpdb

Parse keywords in string based on values contained in an xpdb

parse_title(
  string,
  xpdb,
  problem,
  quiet,
  extra_key = NULL,
  extra_value = NULL,
  ignore_key = NULL
)

Arguments

string

A string containing keywords (e.g. `@nobs`) to be parsed (e.g. title, label, etc.) using values stored in the `xpdb$summary`.

xpdb

An xpose database object.

problem

The $problem number to be used.

quiet

Should messages be displayed to the console.

extra_key

A vector of additional keywords not available in the `xpdb$summary`.

extra_value

A vector of values matching the order of `extra_key`.

ignore_key

A vector of keywords to be ignored i.e. warnings will not be returned.

Value

The parsed `string`.