Display a parameter estimates to the console

Display parameter estimates from an xpdb object to the console.

prm_table(
  xpdb,
  .problem = NULL,
  .subprob = NULL,
  .method = NULL,
  digits = 4,
  transform = TRUE,
  show_all = FALSE
)

Arguments

xpdb
An xpose_data object from which the model output file data will be extracted.

.problem
The problem to be used, by default returns the last one for each file.

.subprob
The subproblem to be used, by default returns the last one for each file.

.method
The estimation method to be used, by default returns the last one for each file

digits
The number of significant digits to be displayed.

transform
Should diagonal OMEGA and SIGMA elements be transformed to standard deviation and off diagonal elements be transformed to correlations.

show_all
Logical, whether the 0 fixed off-diagonal elements should be removed from the output.

See also

Examples

# Store the parameter table
prm <- get_prm(xpdb_ex_pk, .problem = 1)
#> Returning parameter estimates from $prob no.1, subprob no.0, method foce

# Display parameters to the console
prm_table(xpdb_ex_pk, .problem = 1)
#> 
#> Reporting transformed parameters:
#> For the OMEGA and SIGMA matrices, values are reported as standard deviations for the diagonal elements and as correlations for the off-diagonal elements. The relative standard errors (RSE) for OMEGA and SIGMA are reported on the approximate standard deviation scale (SE/variance estimate)/2. Use `transform = FALSE` to report untransformed parameters.
#> 
#> Estimates for $prob no.1, subprob no.0, method foce
#>  Parameter  Label      Value        RSE
#>  THETA1     TVCL       26.29        0.03391
#>  THETA2     TVV        1.348        0.0325
#>  THETA3     TVKA       4.204        0.1925
#>  THETA4     LAG        0.208        0.07554
#>  THETA5     Prop. Err  0.2046       0.1097
#>  THETA6     Add. Err   0.01055      0.3466
#>  THETA7     CRCL on CL 0.007172     0.2366
#>  OMEGA(1,1) IIV CL     0.2701       0.08616
#>  OMEGA(2,2) IIV V      0.195        0.1643
#>  OMEGA(3,3) IIV KA     1.381        0.1463
#>  SIGMA(1,1)            1        fix  -