NONMEM model file parser

Parse NONMEM model files in R format

read_nm_model(file = NULL, runno = NULL, dir = NULL, prefix = "run",
  ext = c(".lst", ".mod", ".ctl", ".txt"), quiet = FALSE)

Arguments

file
Full file name preferably a `.lst` file. Alternative argument to dir, prefix, runno and ext.

runno
Run number to be evaluated.

dir
Location of the model file.

prefix
Prefix of the model file name.

ext
Extension of the model file.

quiet
Logical, if FALSE messages are printed to the console.

Value

A tibble of class model containing the following columns:

problem: a numeric identifier for the $PROBLEM associated with the code.

level: a unique numeric identifier to each subroutine block associated with the code.

subroutine: a character identifier named after the 3 first letters of the subroutine name e.g. "$THETA" and "$TABLE" will become "the" and "tab" respectively. In addtion all output from the .lst is labeled "lst", the general nonmem output e.g. NM-TRAN messages are labeled "oth". With priors thp, tpv, omp, opd, sip, spd abreviations are given to the THETAP, THETAPV, OMEGAP, etc.

code: the code without comments or subroutine names e.g. "$THETA 0.5 ; TVCL" will return 0.5.

comment: the last comment of a record e.g. "0.5 ; Clearance (L/h) ; TVCL" will return "TVCL".

See also

xpose_data, read_nm_tables

Examples

not_run({
  nm_model <- read_nm_model(file = 'run001.lst')
})