Visual predictive checks

Generate visual predictive checks (VPC)

vpc(xpdb, vpc_opt = NULL, mapping = NULL, type = "alp", smooth = TRUE,
  facets = NULL, title = "Visual predictive checks | @run",
  subtitle = "Nsim: @nsim, PI: @vpcpi", caption = "@dir", log = NULL,
  psn_folder = NULL, obs_problem = NULL, sim_problem = NULL, quiet, ...)

Arguments

xpdb
An xpose database object.

vpc_opt
A list of options regarding binning, pi and ci computation. For more information see vpc_opt_set.

mapping
List of aesthetics mappings to be used for the xpose plot (e.g. point_color).

type
String setting the type of plot to be used points 'p', line 'l', area 'a' and text 't' or any combination of the four.

smooth
Should the bins be smoothed (connect bin midpoints, default) or shown as rectangular boxes.

facets
Either a character string to use facet_wrap or a formula to use facet_grid.

title
Plot title. Use NULL to remove.

subtitle
Plot subtitle. Use NULL to remove.

caption
Page caption. Use NULL to remove.

log
String assigning logarithmic scale to axes, can be either '', 'x', y' or 'xy'.

psn_folder
Specify a PsN-generated VPC-folder.

obs_problem
Alternative to the option `psn_folder`. The $problem number to be used for observations. By default returns the last estimation problem.

sim_problem
Alternative to the option `psn_folder`. The $problem number to be used for simulations. By default returns the last simulation problem.

quiet
Logical, if FALSE messages are printed to the console.

...
any additional aesthetics.

Layers mapping

Plots can be customized by mapping arguments to specific layers. The naming convention is layer_option where layer is one of the names defined in the list below and option is any option supported by this layer e.g. point_color = 'blue', smooth_method = 'lm', etc.

point: options to geom_point

line: options to geom_line

guide: options to geom_abline

panel: options to facet_wrap (facets is character) or facet_grid (facets is a formula)

smooth: options to geom_smooth

text: options to geom_text

xscale: options to scale_x_continuous or scale_x_log10

yscale: options to scale_y_continuous or scale_y_log10

Template titles

Template titles can be used to create highly informative diagnostics plots. They can be applied to any plot title, subtitle and caption. Template titles are defined via a single string containing key variables staring with a @ (e.g. @ofv) which will be replaced by their actual value when rendering the plot. For example '@run, @nobs observations in @nind subjects' would become 'run001, 1022 observations in 74 subjects'. The available key variables are listed under template_titles.

See also

vpc

Examples

not_run({
  # Simple vpc using simulated data from the xpdb
  vpc(xpdb_ex_pk)
})