xpose: graphical diagnostics for pharmacometric models

xpose was designed as a ggplot2-based alternative to xpose4. This package aims to reduce the post processing burden and improve diagnostics commonly associated the development of non-linear mixed effect models.

To learn more about xpose, visit our website: https://UUPharmacometrics.github.io/xpose/.

See also

Useful links:

https://github.com/UUPharmacometrics/xpose

Report bugs at https://github.com/UUPharmacometrics/xpose/issues