Import NONMEM output into R

Gather model outputs into a R database

xpose_data(runno = NULL, dir = NULL, file = NULL, prefix = "run",
  ext = ".lst", gg_theme = theme_readable(),
  xp_theme = theme_xp_default(), simtab = NULL, manual_import = NULL,
  skip = NULL, extra_files, quiet, ...)

Arguments

runno
Run number to be evaluated.

dir
Location of the model files.

file
Full file name preferably a `.lst` file. Alternative argument to dir, prefix, runno and ext.

prefix
Prefix of the model file name.

ext
Extension of the model file. Should be one of ".lst" (default), ".out", ".res", ".mod" or ".ctl" for NONMEM.

gg_theme
A ggplot2 theme object (eg. theme_classic).

xp_theme
An xpose theme or vector of modifications to the xpose theme (eg. c(point_color = 'red', line_linetype = 'dashed')).

simtab
If TRUE only reads in simulation tables, if FALSE only reads estimation tables. Default NULL reads all tables. Option not compatible with manual_import.

manual_import
If NULL (default) the names of the output tables to import will be obtained from the model file. To manually import files as in previous versions of xpose, the check the function manual_nm_import.

skip
Character vector be used to skip the import/generation of: 'data', 'files', 'summary' or any combination of the three.

extra_files
A vector of additional output file extensions to be imported. Default is ".ext", ".cov", ".cor", ".phi", ".grd" for NONMEM.

quiet
Logical, if FALSE messages are printed to the console.

...
Additional arguments to be passed to the read_delim functions.

Examples

# NOT RUN {
xpdb <- xpose_data(file = 'run001.lst')
# }