Access model output table data

Access model output table data from an xpdb object.

get_data(xpdb, table = NULL, problem = NULL, quiet)

Arguments

xpdb
An xpose_data object from which the model output file data will be extracted.

table
Name of the output table to be extracted from the xpdb e.g. 'sdtab001'. Alternative to the `problem` argument.

problem
Accesses all tables from the specified problem. Alternative to the `table` argument.

quiet
Logical, if FALSE messages are printed to the console.

Value

By default returns data from the last estimation problem. If only simulation problems are present then the data from last simulation will be returned instead. Object returned as tibble for single tables/problems or a named list for multiple tables/problems.

See also

xpose_data, read_nm_tables

Examples

# By table name
sdtab <- get_data(xpdb_ex_pk, 'sdtab001')
sdtab

#> # A tibble: 1,166 x 19
#>        ID    OCC  DOSE   AMT    SS    II  TIME   TAD   IPRED     CWRES   CPRED
#>    <fctr> <fctr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>     <dbl>   <dbl>
#>  1    110      3   200   100     0     0   0.0   0.0 0.00000  0.000000 0.00000
#>  2    110      3   200     0     0     0   1.0   1.0 0.26359 -0.984100 0.60735
#>  3    110      3   200     0     0     0   1.5   1.5 0.38519 -0.934810 0.77374
#>  4    110      3   200     0     0     0   2.0   2.0 0.44881 -0.058338 0.78518
#>  5    110      3   200     0     0     0   4.0   4.0 0.44686  0.135520 0.45982
#>  6    110      3   200     0     0     0   6.0   6.0 0.34140  0.055376 0.23124
#>  7    110      3   200     0     0     0   8.0   8.0 0.24633  0.122000 0.12925
#>  8    110      8   400   200     1    12  40.0   0.0 0.26111  0.000000 0.10035
#>  9    110      8   400     0     0     0  40.0   0.0 0.26111  0.586710 0.10035
#> 10    110      8   400     0     0     0  40.5   0.5 0.34962 -0.027288 0.38476
#> # ... with 1,156 more rows, and 8 more variables: IWRES <dbl>, EVID <dbl>,
#> #   A1 <dbl>, A2 <dbl>, DV <dbl>, PRED <dbl>, RES <dbl>, WRES <dbl>

# By problem
tables <- get_data(xpdb_ex_pk, problem = 1)
tables

#> # A tibble: 1,166 x 32
#>        ID    SEX   MED1   MED2    OCC  DOSE   AMT    SS    II  TIME   TAD
#>    <fctr> <fctr> <fctr> <fctr> <fctr> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1    110      1      0      1      3   200   100     0     0   0.0   0.0
#>  2    110      1      0      1      3   200     0     0     0   1.0   1.0
#>  3    110      1      0      1      3   200     0     0     0   1.5   1.5
#>  4    110      1      0      1      3   200     0     0     0   2.0   2.0
#>  5    110      1      0      1      3   200     0     0     0   4.0   4.0
#>  6    110      1      0      1      3   200     0     0     0   6.0   6.0
#>  7    110      1      0      1      3   200     0     0     0   8.0   8.0
#>  8    110      1      0      1      8   400   200     1    12  40.0   0.0
#>  9    110      1      0      1      8   400     0     0     0  40.0   0.0
#> 10    110      1      0      1      8   400     0     0     0  40.5   0.5
#> # ... with 1,156 more rows, and 21 more variables: IPRED <dbl>, CWRES <dbl>,
#> #   CPRED <dbl>, IWRES <dbl>, EVID <dbl>, A1 <dbl>, A2 <dbl>, DV <dbl>,
#> #   PRED <dbl>, RES <dbl>, WRES <dbl>, CLCR <dbl>, AGE <dbl>, WT <dbl>,
#> #   KA <dbl>, CL <dbl>, V <dbl>, ALAG1 <dbl>, ETA1 <dbl>, ETA2 <dbl>,
#> #   ETA3 <dbl>

# Tip to list available tables in the xpdb
print(xpdb_ex_pk)

#> run001.lst overview: 
#>  - Software: nonmem 7.3.0 
#>  - Attached files: 
#>    + obs tabs: $prob no.1: catab001.csv, cotab001, patab001, sdtab001 
#>    + sim tabs: $prob no.2: simtab001.zip 
#>    + output files: run001.cor, run001.cov, run001.ext, run001.grd, run001.phi, run001.shk 
#>    + special: <none> 
#>  - gg_theme: theme_readable 
#>  - xp_theme: theme_xp_default 
#>  - Options: dir = analysis/models/pk/, quiet = FALSE, manual_import = NULL