Access model output table data

Access model output table data from an xpdb object.

get_data(xpdb, table = NULL, problem = NULL)

Arguments

xpdb
An xpose_data object from which the model output file data will be extracted.

table
Name of the output table to be extracted from the xpdb e.g. 'sdtab001'. Alternative to the "problem" argument.

problem
Accesses all tables from the specified problem. Alternative to the "table" argument.

Value

A tibble for single file or a named list for multiple files.

See also

xpose_data, read_nm_tables

Examples

# By table name
sdtab <- get_data(xpdb_ex_pk, 'sdtab001')
sdtab

#> # A tibble: 1,166 x 17
#>       ID   OCC  DOSE   AMT    SS    II  TIME   TAD    IPRED      CWRES
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>      <dbl>
#>  1   110     3   200   100     0     0   0.0   0.0 -3.68890  0.0000000
#>  2   110     3   200     0     0     0   1.0   1.0 -1.77630 -2.5234000
#>  3   110     3   200     0     0     0   1.5   1.5 -1.41770 -0.7770900
#>  4   110     3   200     0     0     0   2.0   2.0 -1.22690  0.8963300
#>  5   110     3   200     0     0     0   4.0   4.0 -1.02520  0.4981000
#>  6   110     3   200     0     0     0   6.0   6.0 -1.13610  0.0031566
#>  7   110     3   200     0     0     0   8.0   8.0 -1.35770  0.0379040
#>  8   110     8   400   200     1    12  40.0   0.0 -1.15050  0.0000000
#>  9   110     8   400     0     0     0  40.0   0.0 -1.15050  0.8644100
#> 10   110     8   400     0     0     0  40.5   0.5 -0.83167 -0.5420500
#> # ... with 1,156 more rows, and 7 more variables: CPRED <dbl>, IWRES <dbl>,
#> #   EVID <dbl>, DV <dbl>, PRED <dbl>, RES <dbl>, WRES <dbl>

# By problem
tables <- get_data(xpdb_ex_pk, problem = 1)
tables

#> # A tibble: 1,166 x 30
#>       ID   OCC  DOSE   AMT    SS    II  TIME   TAD    IPRED      CWRES
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl>      <dbl>
#>  1   110     3   200   100     0     0   0.0   0.0 -3.68890  0.0000000
#>  2   110     3   200     0     0     0   1.0   1.0 -1.77630 -2.5234000
#>  3   110     3   200     0     0     0   1.5   1.5 -1.41770 -0.7770900
#>  4   110     3   200     0     0     0   2.0   2.0 -1.22690  0.8963300
#>  5   110     3   200     0     0     0   4.0   4.0 -1.02520  0.4981000
#>  6   110     3   200     0     0     0   6.0   6.0 -1.13610  0.0031566
#>  7   110     3   200     0     0     0   8.0   8.0 -1.35770  0.0379040
#>  8   110     8   400   200     1    12  40.0   0.0 -1.15050  0.0000000
#>  9   110     8   400     0     0     0  40.0   0.0 -1.15050  0.8644100
#> 10   110     8   400     0     0     0  40.5   0.5 -0.83167 -0.5420500
#> # ... with 1,156 more rows, and 20 more variables: CPRED <dbl>, IWRES <dbl>,
#> #   EVID <dbl>, DV <dbl>, PRED <dbl>, RES <dbl>, WRES <dbl>, SEX <dbl>,
#> #   MED1 <dbl>, MED2 <dbl>, CLCR <dbl>, AGE <dbl>, WT <dbl>, KA <dbl>,
#> #   CL <dbl>, V <dbl>, ALAG1 <dbl>, ETA1 <dbl>, ETA2 <dbl>, ETA3 <dbl>

# Tip to list available tables in the xpdb
print(xpdb_ex_pk)

#> run001.lst overview: 
#>  - Software: nonmem 7.3.0 
#>  - Attached files: 
#>    + obs tabs: $prob no.1: catab001.csv, cotab001, patab001, sdtab001 
#>    + sim tabs: $prob no.2: simtab001.zip 
#>    + output files: run001.cor, run001.cov, run001.ext, run001.grd, run001.phi, run001.shk 
#>    + special: <none> 
#>  - gg_theme: theme_readable 
#>  - xp_theme: theme_xp_default 
#>  - Options: dir = analysis/models/pk/, quiet = TRUE, manual_import = NULL