Access model output file data

Access model output file data from an xpdb object.

get_file(xpdb, file = NULL, ext = NULL, problem = NULL, subprob = NULL,
  quiet = FALSE)

Arguments

xpdb
An xpose_data object from which the model output file data will be extracted.

file
Full name of the file to be extracted from the xpdb e.g. 'run001.phi'. Alternative to the 'ext' argument.

ext
Extension of the file to be extracted from the xpdb e.g. 'phi'. Alternative to the 'file' argument.

problem
The problem to be used, by default returns the last one for each file.

subprob
The subproblem to be used, by default returns the last one for each file.

quiet
Logical, if FALSE messages are printed to the console.

Value

A tibble for single file or a named list for multiple files.

See also

xpose_data, read_nm_files

Examples

# Single file (returns a tibble)
ext_file <- get_file(xpdb_ex_pk, file = 'run001.ext')

#> Returning data from problem no.1 sub-problem no.0

ext_file

#> # A tibble: 34 x 15
#>    ITERATION  THETA1  THETA2  THETA3   THETA4   THETA5     THETA6 SIGMA.1.1.
#>        <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>      <dbl>      <dbl>
#>  1         0 27.5000 1.56500 2.10000 0.254000 0.230000 0.00800000          1
#>  2         1 27.4212 1.56156 2.10527 0.259727 0.419087 0.00797891          1
#>  3         2 27.3474 1.56221 2.10859 0.261258 0.451587 0.00797180          1
#>  4         3 23.1413 1.67880 2.33794 0.335797 0.436782 0.00763214          1
#>  5         4 28.7697 1.54124 2.53932 0.381648 0.429614 0.00745381          1
#>  6         5 28.3703 1.47870 2.67525 0.385315 0.428860 0.00746121          1
#>  7         6 28.5457 1.53846 3.54898 0.401145 0.426591 0.00750659          1
#>  8         7 29.2138 1.67721 8.44612 0.206586 0.432386 0.00739348          1
#>  9         8 28.5851 1.64530 9.55926 0.228455 0.433989 0.00652994          1
#> 10         9 28.1226 1.63433 8.01904 0.228860 0.432755 0.00634526          1
#> # ... with 24 more rows, and 7 more variables: OMEGA.1.1. <dbl>,
#> #   OMEGA.2.1. <dbl>, OMEGA.2.2. <dbl>, OMEGA.3.1. <dbl>, OMEGA.3.2. <dbl>,
#> #   OMEGA.3.3. <dbl>, OBJ <dbl>

# Multiple files (returns a list)
files <- get_file(xpdb_ex_pk, file = c('run001.ext', 'run001.phi'))
files

#> $run001.ext
#> # A tibble: 34 x 15
#>    ITERATION  THETA1  THETA2  THETA3   THETA4   THETA5     THETA6 SIGMA.1.1.
#>        <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>      <dbl>      <dbl>
#>  1         0 27.5000 1.56500 2.10000 0.254000 0.230000 0.00800000          1
#>  2         1 27.4212 1.56156 2.10527 0.259727 0.419087 0.00797891          1
#>  3         2 27.3474 1.56221 2.10859 0.261258 0.451587 0.00797180          1
#>  4         3 23.1413 1.67880 2.33794 0.335797 0.436782 0.00763214          1
#>  5         4 28.7697 1.54124 2.53932 0.381648 0.429614 0.00745381          1
#>  6         5 28.3703 1.47870 2.67525 0.385315 0.428860 0.00746121          1
#>  7         6 28.5457 1.53846 3.54898 0.401145 0.426591 0.00750659          1
#>  8         7 29.2138 1.67721 8.44612 0.206586 0.432386 0.00739348          1
#>  9         8 28.5851 1.64530 9.55926 0.228455 0.433989 0.00652994          1
#> 10         9 28.1226 1.63433 8.01904 0.228860 0.432755 0.00634526          1
#> # ... with 24 more rows, and 7 more variables: OMEGA.1.1. <dbl>,
#> #   OMEGA.2.1. <dbl>, OMEGA.2.2. <dbl>, OMEGA.3.1. <dbl>, OMEGA.3.2. <dbl>,
#> #   OMEGA.3.3. <dbl>, OBJ <dbl>
#> 
#> $run001.phi
#> # A tibble: 74 x 12
#>    SUBJECT_NO    ID     ETA.1.      ETA.2.    ETA.3.   ETC.1.1.    ETC.2.1.
#>         <dbl> <dbl>      <dbl>       <dbl>     <dbl>      <dbl>       <dbl>
#>  1          1   110 -0.0458450 -0.00353128 -2.146040 0.00989773 0.000203327
#>  2          2   112 -0.1051620  0.02012950 -0.162595 0.00710118 0.002104120
#>  3          3   113 -0.0461785  0.06741830  2.488430 0.00713534 0.002342970
#>  4          4   121  0.1653850  0.08542600  2.707560 0.00566912 0.003772380
#>  5          5   123  0.0106041 -0.04118790  2.150870 0.01404410 0.001713300
#>  6          6   124  0.2025370  0.00406404 -1.172320 0.00620707 0.002575780
#>  7          7   126  0.0385193  0.02425710  3.265320 0.00675844 0.002599590
#>  8          8   127  0.1901370 -0.03114860  3.107180 0.00588325 0.003490500
#>  9          9   128  0.0268498  0.00984491 -1.382470 0.00656982 0.002521020
#> 10         10   210 -0.3111150 -0.03701740 -0.366683 0.00796322 0.001761540
#> # ... with 64 more rows, and 5 more variables: ETC.2.2. <dbl>, ETC.3.1. <dbl>,
#> #   ETC.3.2. <dbl>, ETC.3.3. <dbl>, OBJ <dbl>
#> 

# Tip to list available files in the xpdb
print(xpdb_ex_pk)

#> run001.lst overview: 
#>  - Software: nonmem 7.3.0 
#>  - Attached files: 
#>    + tables: $prob no.1: catab001.csv, cotab001, patab001, sdtab001 
#>    + sim tabs: $prob no.2: simtab001.zip 
#>    + output files: run001.cor, run001.cov, run001.ext, run001.grd, run001.phi, run001.shk 
#>  - gg_theme: theme_readable 
#>  - xp_theme: theme_xp_default 
#>  - Options: dir = analysis/models/pk/, quiet = TRUE, manual_import = NULL