Access model output file data

Access model output file data from an xpdb object.

get_file(xpdb, file = NULL, ext = NULL, problem = NULL, subprob = NULL,
  method = NULL, quiet)

Arguments

xpdb
An xpose_data object from which the model output file data will be extracted.

file
Full name of the file to be extracted from the xpdb e.g. 'run001.phi'. Alternative to the 'ext' argument.

ext
Extension of the file to be extracted from the xpdb e.g. 'phi'. Alternative to the 'file' argument.

problem
The problem to be used, by default returns the last one for each file.

subprob
The subproblem to be used, by default returns the last one for each file.

method
The estimation method to be used (e.g. 'foce', 'imp', 'saem'), by default returns the last one for each file.

quiet
Logical, if FALSE messages are printed to the console.

Value

A tibble for single file or a named list for multiple files.

See also

xpose_data, read_nm_files

Examples

# Single file (returns a tibble)
ext_file <- get_file(xpdb_ex_pk, file = 'run001.ext')

#> Returning data from run001.ext, $prob no.1, subprob no.0, method foce

ext_file

#> # A tibble: 30 x 15
#>    ITERATION  THETA1  THETA2  THETA3   THETA4   THETA5     THETA6 `SIGMA(1,1)`
#>        <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>      <dbl>        <dbl>
#>  1         0 27.5000 1.56500 2.10000 0.254000 0.230000 0.00800000            1
#>  2         1 27.1315 1.50371 2.19470 0.263436 0.310468 0.00799037            1
#>  3         2 24.0386 1.05681 3.37071 0.341244 0.287332 0.00792983            1
#>  4         3 26.6619 1.23811 4.01449 0.409304 0.279560 0.00790061            1
#>  5         4 24.0386 1.44520 4.43650 0.400794 0.277556 0.00794055            1
#>  6         5 23.9144 1.44930 4.52519 0.404786 0.277363 0.00796163            1
#>  7         6 24.7365 1.50060 7.71427 0.410008 0.274756 0.00848706            1
#>  8         7 24.6300 1.49149 8.80826 0.409741 0.273132 0.00859991            1
#>  9         8 24.5996 1.47520 8.44218 0.413982 0.269094 0.00794411            1
#> 10         9 24.6701 1.45429 7.04078 0.413384 0.259226 0.00574881            1
#> # ... with 20 more rows, and 7 more variables: `OMEGA(1,1)` <dbl>,
#> #   `OMEGA(2,1)` <dbl>, `OMEGA(2,2)` <dbl>, `OMEGA(3,1)` <dbl>,
#> #   `OMEGA(3,2)` <dbl>, `OMEGA(3,3)` <dbl>, OBJ <dbl>

# Multiple files (returns a list)
files <- get_file(xpdb_ex_pk, file = c('run001.ext', 'run001.phi'))

#> Returning data from run001.ext, run001.phi

files

#> $run001.ext_prob_1_subprob_0_foce
#> # A tibble: 30 x 15
#>    ITERATION  THETA1  THETA2  THETA3   THETA4   THETA5     THETA6 `SIGMA(1,1)`
#>        <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>      <dbl>        <dbl>
#>  1         0 27.5000 1.56500 2.10000 0.254000 0.230000 0.00800000            1
#>  2         1 27.1315 1.50371 2.19470 0.263436 0.310468 0.00799037            1
#>  3         2 24.0386 1.05681 3.37071 0.341244 0.287332 0.00792983            1
#>  4         3 26.6619 1.23811 4.01449 0.409304 0.279560 0.00790061            1
#>  5         4 24.0386 1.44520 4.43650 0.400794 0.277556 0.00794055            1
#>  6         5 23.9144 1.44930 4.52519 0.404786 0.277363 0.00796163            1
#>  7         6 24.7365 1.50060 7.71427 0.410008 0.274756 0.00848706            1
#>  8         7 24.6300 1.49149 8.80826 0.409741 0.273132 0.00859991            1
#>  9         8 24.5996 1.47520 8.44218 0.413982 0.269094 0.00794411            1
#> 10         9 24.6701 1.45429 7.04078 0.413384 0.259226 0.00574881            1
#> # ... with 20 more rows, and 7 more variables: `OMEGA(1,1)` <dbl>,
#> #   `OMEGA(2,1)` <dbl>, `OMEGA(2,2)` <dbl>, `OMEGA(3,1)` <dbl>,
#> #   `OMEGA(3,2)` <dbl>, `OMEGA(3,3)` <dbl>, OBJ <dbl>
#> 
#> $run001.phi_prob_1_subprob_0_foce
#> # A tibble: 74 x 12
#>    SUBJECT_NO    ID   `ETA(1)`   `ETA(2)`    `ETA(3)` `ETC(1,1)`   `ETC(2,1)`
#>         <dbl> <dbl>      <dbl>      <dbl>       <dbl>      <dbl>        <dbl>
#>  1          1   110 -0.0806742 -0.0607414 -1.76376000 0.01277110 -0.000839822
#>  2          2   112 -0.0569953 -0.0759594 -1.12982000 0.00522566 -0.001305390
#>  3          3   113 -0.0350773  0.0919003  0.00892949 0.02825720 -0.004718640
#>  4          4   121  0.1467100  0.5133580  0.48995100 0.01083650 -0.003341270
#>  5          5   123  0.0212695 -0.0755525  0.33757400 0.03400030 -0.003807550
#>  6          6   124  0.1261170  0.1555090 -1.09301000 0.02349470 -0.003032780
#>  7          7   126 -0.1294000  0.5399580  0.45971900 0.00546046 -0.002342990
#>  8          8   127 -0.0031396  0.3230990  0.39069200 0.01973200 -0.004476720
#>  9          9   128  0.0604116  0.0220288 -1.48978000 0.02527590 -0.001712680
#> 10         10   210 -0.3465460 -0.4536260 -0.90240000 0.00768758 -0.002431600
#> # ... with 64 more rows, and 5 more variables: `ETC(2,2)` <dbl>,
#> #   `ETC(3,1)` <dbl>, `ETC(3,2)` <dbl>, `ETC(3,3)` <dbl>, OBJ <dbl>
#> 

# Tip to list available files in the xpdb
print(xpdb_ex_pk)

#> run001.lst overview: 
#>  - Software: nonmem 7.3.0 
#>  - Attached files: 
#>    + obs tabs: $prob no.1: catab001.csv, cotab001, patab001, sdtab001 
#>    + sim tabs: $prob no.2: simtab001.zip 
#>    + output files: run001.cor, run001.cov, run001.ext, run001.grd, run001.phi, run001.shk 
#>    + special: <none> 
#>  - gg_theme: theme_readable 
#>  - xp_theme: theme_xp_default 
#>  - Options: dir = analysis/models/pk/, quiet = FALSE, manual_import = NULL