Access model summary data

Access model summary data from an xpdb object.

get_summary(xpdb, problem = NULL, subprob = NULL, only_last = FALSE)

Arguments

xpdb
An xpose_data object from which the summary data will be extracted.

problem
The problem to be used, by default returns the last one for each label.

subprob
The subproblem to be used, by default returns the last one for each label.

only_last
Logical, if TRUE only the last record for each label is returned in case of multiple problem and/or subproblem. If FALSE all values are returned.

Value

A tibble of model summary.

See also

xpose_data, template_titles

Examples

run_summary <- get_summary(xpdb_ex_pk)
run_summary

#> # A tibble: 28 x 5
#>    problem subprob                 descr       label               value
#>      <dbl>   <dbl>                 <chr>       <chr>               <chr>
#>  1       0       0       Run description       descr    Run for testthat
#>  2       0       0         Run directory         dir analysis/models/pk/
#>  3       0       0            Run errors      errors                  na
#>  4       0       0   ESAMPLE seed number esampleseed                  na
#>  5       0       0              Run file        file          run001.lst
#>  6       0       0     Number of ESAMPLE    nesample                  na
#>  7       0       0 Number of simulations        nsim                  na
#>  8       0       0       Reference model         ref                 000
#>  9       0       0            Run number         run              run001
#> 10       0       0       Simulation seed     simseed                  na
#> # ... with 18 more rows