Manually define nonmem tables to be imported

Manually provide names of the table files to be imported by xpose_data.

manual_nm_import(
  tab_names = c("sdtab", "mutab", "patab", "catab", "cotab", "mytab", "extra", "xptab",
    "cwtab"),
  tab_suffix = "",
  sim_suffix = "sim"
)

Arguments

tab_names

Provide the name of the tables to import e.g. 'sdtab', 'patab', 'cotab', 'catab' for NONMEM.

tab_suffix

Default is '', but can be changed to any character string to be used as suffix in the table names.

sim_suffix

Default is 'sim', but can be changed to any character string to be used as suffix in the simulation table names e.g. sdtab001sim.

Details

In order to be imported manually, table names must follow the following convention: <tab_names><runno><tab/sim_suffix> e.g. sdtab001sim. When the argument `file` is used in xpose_data, the <runno> part is guessed by taking the portion of the string starting by any digit and ending at the file extension e.g. file = run001a.mod will guess <runno> as `001a`. If no valid <runno> can be guessed, xpose will return an error. In this case it is advised to use the xpose_data argument `runno` directly rather than `file` hence preventing xpose from having to guess <runno>.

Note that with manual table import xpose still reads in the NONMEM model file in order to generate the run summary.

See also

Examples

if (FALSE) {
# Import all names specified by default as in xpose4
xpose_data(runno = '001', manual_import = manual_nm_import())

# Import a specific table name
xpose_data(runno = '001', manual_import = manual_nm_import(tab_names = 'mytab'))
}