Generate a list of options for VPC data generation

Provide a list of options to vpc_data function.

vpc_opt(
  bins = "jenks",
  n_bins = "auto",
  bin_mid = "mean",
  pred_corr = FALSE,
  pred_corr_lower_bnd = 0,
  pi = c(0.025, 0.975),
  ci = c(0.025, 0.975),
  lloq = NULL,
  uloq = NULL,
  rtte = FALSE,
  rtte_calc_diff = TRUE,
  events = NULL,
  kmmc = NULL,
  reverse_prob = FALSE,
  as_percentage = TRUE
)

Arguments

bins

Binning method, can be one of 'density', 'time', 'data', 'none', or one of the approaches available in classInterval() such as 'jenks' (default), 'pretty', or a numeric vector specifying the bin separators.

n_bins

When using the 'auto' binning method, what number of bins to aim for.

bin_mid

Specify how to is the mid bin value calculated, can be either 'mean' for the mean of all timepoints (default) or 'middle' to use the average of the bin boundaries.

pred_corr

Option reserved to continuous VPC. Logical, should a prediction correction (pcVPC) of the data be used.

pred_corr_lower_bnd

Option reserved to continuous VPC. Lower bound for the prediction-correction.

pi

Option reserved to continuous VPC. Simulated prediction interval to plot. Default is c(0.05, 0.95).

ci

Confidence interval around the percentiles to plot. Default is c(0.05, 0.95)

lloq

Number or NULL indicating lower limit of quantification. Default is NULL.

uloq

Number or NULL indicating upper limit of quantification. Default is NULL.

rtte

Option reserved to time-to-event VPC. Is the data repeated time-to-event (RTTE) TRUE or single time-to-event (TTE) FALSE.

rtte_calc_diff

Option reserved to time-to-event VPC. Should the time be recalculated? When simulating in NONMEM, you will probably need to set this to TRUE to recalculate the TIME to the relative time between events (unless you output the time difference between events and specify that as independent variable in the index.

events

Option reserved to time-to-event VPC. Numeric vector describing which events to show a VPC for when repeated TTE data, e.g. c(1:4). Default is NULL, which shows all events.

kmmc

Option reserved to time-to-event VPC. Either NULL for regular TTE VPC (default), or a variable name for a KMMC plot (e.g. 'WT').

reverse_prob

Option reserved to time-to-event VPC. Should the probability be reversed (i.e. plot 1-probability).

as_percentage

Should the Y-scale be in percent (0-100) TRUE (default), or standard (0-1) FALSE.

See also

Examples

vpc_opt()
#> $bins
#> [1] "jenks"
#> 
#> $n_bins
#> [1] "auto"
#> 
#> $bin_mid
#> [1] "mean"
#> 
#> $pred_corr
#> [1] FALSE
#> 
#> $pred_corr_lower_bnd
#> [1] 0
#> 
#> $pi
#> [1] 0.025 0.975
#> 
#> $ci
#> [1] 0.025 0.975
#> 
#> $lloq
#> NULL
#> 
#> $uloq
#> NULL
#> 
#> $rtte
#> [1] FALSE
#> 
#> $rtte_calc_diff
#> [1] TRUE
#> 
#> $events
#> NULL
#> 
#> $kmmc
#> NULL
#> 
#> $reverse_prob
#> [1] FALSE
#> 
#> $as_percentage
#> [1] TRUE
#> 
#> $usr_call
#> NULL
#>